requirements.txt
asgiref==3.5.0
astroid==2.9.3
black==22.1.0
click==8.0.3
colorama==0.4.4
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pylint==2.12.2
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
tzdata==2021.5
wrapt==1.13.3
